use usize rather than u64 - I prefer the native machine word size - the numbers do not need 64 bit yet?

+++ rust/pe001/src/main.rs
@@ -1,6 +1,6 @@
-fn pe001() -> u64 {
-    let imax:u64 = 1000;
-    let mut total:u64 = 0;
+fn pe001() -> usize {
+    let imax:usize = 1000;
+    let mut total:usize = 0;
 
     for i in 0 .. imax {
         if i % 3 == 0 || i % 5 == 0 {
